Create recipe-assistant.py
created recipe-assistant.py.
this will be the entire program.
created the framework for the program.
added all the functions i think will be necessary, but a lot are empty for now.
function that allows user to input ingredients in functional i believe.
main menu created and is functional.
documentation and comments within code.
current issue to think about: how am i gonna associate ingredients with recipes? and am i able to then further associate steps of the recipe with the recipe, or should i scratch that idea and have this program just suggest names of recipes?