updated recipe-assistant.py
created links to (unfinished) ingredients and recipe .txt files. opened and read these files to convert them to lists within the program. this will be used later in the program to check if the ingredients users are entering are accepted by the program, and to suggest recipes based on user ingredients.
bug in the add_ingredients function! not sure why but will be solved later.